requirements.txt
praw>=7.7.0
python-dotenv>=1.1.1